Handled navigation from walkthrough screens to welcome/auth screen
