+++ app/src/main/java/gads/mobile/ecom05/fragments/OnBoardingScreenFragment.kt
@@ -8,6 +8,8 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.RelativeLayout
 import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.fragment.findNavController
 import com.google.android.material.tabs.TabLayoutMediator
 import gads.mobile.ecom05.R
 import gads.mobile.ecom05.adapters.WalkThroughScreenPagerAdapter
@@ -16,6 +18,7 @@ import kotlinx.android.synthetic.main.fragment_onboarding_screen.view.*
 import java.util.prefs.Preferences
 const val ONBOARDING_PREF="gads.mobile.ecom05.onboarding"
 class OnBoardingScreenFragment: Fragment() {
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -29,22 +32,23 @@ class OnBoardingScreenFragment: Fragment() {
         }.attach()
         view.button_get_started.setOnClickListener {
             onBoardingFinished()
-            addWelcomeScreenToContainer()
+            NavHostFragment.findNavController(this)
+                .navigate(R.id.action_onBoardingScreenFragment_to_welcomeAuthPromptFragment)
         }
         return view
 
     }
-
-    private fun addWelcomeScreenToContainer() {
-      val welcomeAuthPromptFragment=WelcomeAuthPromptFragment()
-        val fragmentTransaction= fragmentManager?.beginTransaction()
-        val container= requireActivity().container.id
-
-            fragmentTransaction?.replace(container,welcomeAuthPromptFragment)
-        fragmentTransaction?.addToBackStack(null)
-        fragmentTransaction?.commit()
-
-    }
+//This code block is replaced with navigation graph
+//    private fun addWelcomeScreenToContainer() {
+//      val welcomeAuthPromptFragment=WelcomeAuthPromptFragment()
+//        val fragmentTransaction= fragmentManager?.beginTransaction()
+//        val container= requireActivity().container.id
+//
+//            fragmentTransaction?.replace(container,welcomeAuthPromptFragment)
+//        fragmentTransaction?.addToBackStack(null)
+//        fragmentTransaction?.commit()
+//
+//    }
 
     private fun onBoardingFinished() {
         val sharePreferences=requireActivity().getSharedPreferences(ONBOARDING_PREF, Context.MODE_PRIVATE)
