+++ app/src/main/res/navigation/welcome_auth_navigation.xml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/welcome_auth_navigation"
+    app:startDestination="@id/loginFragment2">
+
+    <fragment
+        android:id="@+id/loginFragment2"
+        android:name="gads.mobile.ecom05.fragments.login.LoginFragment"
+        android:label="fragment_login"
+        tools:layout="@layout/fragment_login" >
+        <action
+            android:id="@+id/action_loginFragment2_to_mainActivity"
+            app:destination="@id/mainActivity" />
+    </fragment>
+    <fragment
+        android:id="@+id/signupFragment2"
+        android:name="gads.mobile.ecom05.fragments.signup.SignupFragment"
+        android:label="fragment_signup"
+        tools:layout="@layout/fragment_signup" >
+        <action
+            android:id="@+id/action_signupFragment2_to_mainActivity"
+            app:destination="@id/mainActivity" />
+    </fragment>
+    <fragment
+        android:id="@+id/welcomeAuthPromptFragment"
+        android:name="gads.mobile.ecom05.fragments.WelcomeAuthPromptFragment"
+        android:label="WelcomeAuthPromptFragment" >
+        <action
+            android:id="@+id/action_welcomeAuthPromptFragment_to_signupFragment2"
+            app:destination="@id/signupFragment2" />
+        <action
+            android:id="@+id/action_welcomeAuthPromptFragment_to_loginFragment2"
+            app:destination="@id/loginFragment2" />
+    </fragment>
+    <fragment
+        android:id="@+id/onBoardingScreenFragment"
+        android:name="gads.mobile.ecom05.fragments.walkthrough.OnBoardingScreenFragment"
+        android:label="OnBoardingScreenFragment" >
+        <action
+            android:id="@+id/action_onBoardingScreenFragment_to_welcomeAuthPromptFragment"
+            app:destination="@id/welcomeAuthPromptFragment" />
+    </fragment>
+    <activity
+        android:id="@+id/mainActivity"
+        android:name="gads.mobile.ecom05.MainActivity"
+        android:label="activity_main"
+        tools:layout="@layout/activity_main" />
+</navigation>
